+++ cli/cli_test.go
@@ -17,87 +17,91 @@ func TestParseFlags(t *testing.T) {
 	}{
 		{
 			args:  []string{},
-			flags: &Flags{Color: true},
+			flags: &Flags{Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-H", "some-host"},
-			flags: &Flags{Host: "some-host", Color: true},
+			flags: &Flags{Host: "some-host", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--host", "some-host"},
-			flags: &Flags{Host: "some-host", Color: true},
+			flags: &Flags{Host: "some-host", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-M", "some-machine"},
-			flags: &Flags{Machine: "some-machine", Color: true},
+			flags: &Flags{Machine: "some-machine", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--machine", "some-machine"},
-			flags: &Flags{Machine: "some-machine", Color: true},
+			flags: &Flags{Machine: "some-machine", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-K", "some-instance"},
-			flags: &Flags{Kitchen: "some-instance", Color: true},
+			flags: &Flags{Kitchen: "some-instance", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--kitchen", "some-instance"},
-			flags: &Flags{Kitchen: "some-instance", Color: true},
+			flags: &Flags{Kitchen: "some-instance", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--ssh", "x=1", "--ssh", "y 2 3"},
-			flags: &Flags{SSHOptions: []string{"x=1", "y 2 3"}, Color: true},
+			flags: &Flags{SSHOptions: []string{"x=1", "y 2 3"}, Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--rsync", "-x", "--rsync", "--y"},
-			flags: &Flags{RsyncOptions: []string{"-x", "--y"}, Color: true},
+			flags: &Flags{RsyncOptions: []string{"-x", "--y"}, Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--resolver", "some-resolver"},
-			flags: &Flags{Resolver: "some-resolver", Color: true},
+			flags: &Flags{Resolver: "some-resolver", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-i", "1.2.3"},
-			flags: &Flags{ChefVersion: "1.2.3", Color: true},
+			flags: &Flags{ChefVersion: "1.2.3", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--install-chef", "1.2.3"},
-			flags: &Flags{ChefVersion: "1.2.3", Color: true},
+			flags: &Flags{ChefVersion: "1.2.3", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-F", "some-format"},
-			flags: &Flags{Format: "some-format", Color: true},
+			flags: &Flags{Format: "some-format", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--format", "some-format"},
-			flags: &Flags{Format: "some-format", Color: true},
+			flags: &Flags{Format: "some-format", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-l", "some-level"},
-			flags: &Flags{LogLevel: "some-level", Color: true},
+			flags: &Flags{LogLevel: "some-level", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--log_level", "some-level"},
-			flags: &Flags{LogLevel: "some-level", Color: true},
+			flags: &Flags{LogLevel: "some-level", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"-j", "some-file"},
-			flags: &Flags{JSONFile: "some-file", Color: true},
+			flags: &Flags{JSONFile: "some-file", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"--json-attributes", "some-file"},
-			flags: &Flags{JSONFile: "some-file", Color: true},
+			flags: &Flags{JSONFile: "some-file", Sudo: true, Color: true},
 		},
 		{
 			args:  []string{"some-recipe", "another-recipe"},
-			flags: &Flags{Recipes: []string{"some-recipe", "another-recipe"}, Color: true},
+			flags: &Flags{Recipes: []string{"some-recipe", "another-recipe"}, Sudo: true, Color: true},
+		},
+		{
+			args:  []string{"--sudo=false"},
+			flags: &Flags{Sudo: false, Color: true},
 		},
 		{
 			args:  []string{"--color=false"},
-			flags: &Flags{Color: false},
+			flags: &Flags{Sudo: true, Color: false},
 		},
 		{
 			args:  []string{"--version"},
-			flags: &Flags{ShowVersion: true, Color: true},
+			flags: &Flags{ShowVersion: true, Sudo: true, Color: true},
 		},
 		{
 			args: []string{"--machine", "some-machine", "-l", "some-level", "-i", "true", "some-recipe"},
@@ -106,6 +110,7 @@ func TestParseFlags(t *testing.T) {
 				ChefVersion: "true",
 				LogLevel:    "some-level",
 				Recipes:     []string{"some-recipe"},
+				Sudo:        true,
 				Color:       true,
 			},
 		},